Rename cards.ts to cardefs.ts
